Clarify `eslint --print-config` usage
I thought that `--print-config` to a config file path as argument, but
it takes _any_ file path as argument. ESLint starts looking for config
files from that file and upwards in the file system.

(If one wishes to specify a config file via a flag, the command is
`eslint --config config.js --print-config some/file.js`.)